requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
pycparser==2.21
pygit2==1.11.1
requests==2.28.1
smmap==5.0.0
urllib3==1.26.13
utils==1.0.1
